Reorganize parsing subpackage
Move custom param types into a subpackage and simplify/combine the
command decorators into `globus_cli.parsing.commands`
