+++ src/globus_cli/parsing/param_types/__init__.py
@@ -1,0 +1,17 @@
+from .endpoint_plus_path import (
+    ENDPOINT_PLUS_OPTPATH,
+    ENDPOINT_PLUS_REQPATH,
+    EndpointPlusPath,
+)
+from .identity_type import IdentityType
+from .location import LocationType
+from .task_path import TaskPath
+
+__all__ = (
+    "ENDPOINT_PLUS_OPTPATH",
+    "ENDPOINT_PLUS_REQPATH",
+    "EndpointPlusPath",
+    "IdentityType",
+    "LocationType",
+    "TaskPath",
+)
